pyproject.toml
[project]
name = "jortt-duck"
version = "0.1.0"
description = "Jortt API to MotherDuck data pipeline using dlt"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "dlt[duckdb]>=1.20.0",
    "marimo>=0.18.4",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "pyzmq>=27.1.0",
    "requests>=2.32.5",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
